NEWS: update.
Added the 2 new PDB functions. I should not have added these so close to release
(because these add 2 translatable strings for undo), but anyway the added
strings are barely GUI-facing as these functions are for plug-in developers.

And I really wanted to finally review these MRs which have been sitting there
for months. Sorry translators!