app, meson: fix single window test failure
Our CI meson tests for single window failed even though the sessionrc
looked the same as the one created by the autotools tests.

Looking at the Makefile.am it explicitly states that the tests shouldn't
be run in parallel or in a different order than specified, which was not
the case in our meson build.

We change this by setting is_parallel to false for each build, and
setting a different priority for each test. Since higher priority values
execute first we count down.