app, libgimp, pdb: use 64-bit window id
This is a follow-up of the clang issues signaled in #8649. When testing
CLANG64 on Windows under MSYS2, there was another similar issue in
gimpwidgets-utils.c.
In addition to that, the previous cast using GPOINTER_TO_INT proved
incorrect, since that discards anything except the lower 32-bits.
Since we already return guint64 here, we will cast this to guint64 too.

By changing the return to guint64 in gimpwidgets-utils.c too, we kind of
snowball into needing to change multiple functions that so far used
guint32 for window ids.