app: make GimpActionGroup a subclass of GActionGroup.
It should make it possible to attach an action group to a widget, instead of
making it global to the application. This will be useful/used for context
actions, such as actions which are only relevant to a dock and which will
therefore be duplicated to each dock (with the specific dock as user action so
that we work on the right part of the GUI).

I have not implemented all virtual methods. Not sure if they are all necessary.